# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: Cabales API
  version: 1.2.0
  license: { name: ISC, identifier: ISC }
  description: >-
    API de autenticación, grupos, eventos, gastos exactos, liquidaciones, fondos, presupuestos, documentos, OCR, Cabudas, estadísticas, avisos, logros y privacidad. Importes en centavos.
    Todas las respuestas /api/v1 usan Cache-Control private, no-store; health y ready usan no-store.
servers:
  - url: /
...
paths:
...
  /api/v1/groups:
    get:
      tags: [Groups]
      summary: Lista los grupos del actor
      operationId: listGroups
      security: [{ sessionCookie: [] }]
      responses:
        '200': { $ref: '#/components/responses/GroupListSuccess' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
    post:
      tags: [Groups]
      summary: Crea grupo y membresía OWNER
      operationId: createGroup
      security: [{ sessionCookie: [], csrfCookie: [], csrfHeader: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/GroupInput' }
      responses:
        '201': { $ref: '#/components/responses/GroupSuccess' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /api/v1/groups/{groupId}:
    parameters:
      - $ref: '#/components/parameters/GroupId'
    get:
      tags: [Groups]
      summary: Obtiene grupo y miembros
      operationId: getGroup
      security: [{ sessionCookie: [] }]
      responses:
        '200': { $ref: '#/components/responses/GroupDetailSuccess' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
    patch:
      tags: [Groups]
      summary: Actualiza grupo como OWNER o ADMIN
      operationId: updateGroup
      security: [{ sessionCookie: [], csrfCookie: [], csrfHeader: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/GroupPatch' }
      responses:
        '200': { $ref: '#/components/responses/GroupSuccess' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
    delete:
      tags: [Groups]
      summary: Elimina un grupo vacío como OWNER
      operationId: deleteGroup
      security: [{ sessionCookie: [], csrfCookie: [], csrfHeader: [] }]
      responses:
        '200': { $ref: '#/components/responses/MutationSuccess' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /api/v1/groups/{groupId}/invitations:
    get:
      tags: [Groups]
      summary: 'Lista invitaciones del grupo (propietario o administrador)'
      operationId: listInvitations
      parameters:
        - $ref: '#/components/parameters/GroupId'
        - name: status
          in: query
          required: false
          schema:
            enum: [PENDING, ACCEPTED, REVOKED, EXPIRED]
      security:
        - sessionCookie: []
      responses:
        '200':
          description: 'Operación correcta'
          content:
            application/json:
              schema:
                type: object
                required: [success, data]
                properties:
                  success:
                    const: true
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Invitation' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
    post:
      tags: [Groups]
      summary: Crea una invitación y devuelve su token una sola vez
      operationId: createInvitation
      parameters:
        - $ref: '#/components/parameters/GroupId'
      security: [{ sessionCookie: [], csrfCookie: [], csrfHeader: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/InvitationInput' }
      responses:
        '201': { $ref: '#/components/responses/InvitationSuccess' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /api/v1/groups/invitations/accept:
    post:
      tags: [Groups]
      summary: Acepta una invitación ligada al correo autenticado
      operationId: acceptInvitation
      security: [{ sessionCookie: [], csrfCookie: [], csrfHeader: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AcceptInvitationInput' }
      responses:
        '200': { $ref: '#/components/responses/MembershipSuccess' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '410': { $ref: '#/components/responses/Gone' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
...
  /api/v1/groups/{groupId}/invitations/{invitationId}/resend:
    parameters:
      - $ref: '#/components/parameters/GroupId'
      - $ref: '#/components/parameters/InvitationId'
    post:
      tags: [Groups]
      summary: 'Reenvía rotando el token; reabre invitaciones vencidas'
      operationId: resendInvitation
      security:
        - sessionCookie: []
          csrfCookie: []
          csrfHeader: []
      responses:
        '200':
          description: 'Operación correcta'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InvitationEnvelope' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /api/v1/groups/{groupId}/invitations/{invitationId}/revoke:
    parameters:
      - $ref: '#/components/parameters/GroupId'
      - $ref: '#/components/parameters/InvitationId'
    post:
      tags: [Groups]
      summary: 'Revoca una invitación pendiente'
      operationId: revokeInvitation
      security:
        - sessionCookie: []
          csrfCookie: []
          csrfHeader: []
      responses:
        '200':
          description: 'Operación correcta'
          content:
            application/json:
              schema:
                type: object
                required: [success, data]
                properties:
                  success:
                    const: true
                  data: { $ref: '#/components/schemas/Invitation' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /api/v1/groups/invitations/preview:
    post:
      tags: [Groups]
      summary: 'Previsualiza una invitación sin aceptarla'
      operationId: previewInvitation
      description: 'No revela el correo invitado; solo indica si coincide con la sesión.'
      security:
        - sessionCookie: []
          csrfCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AcceptInvitationInput' }
      responses:
        '200':
          description: 'Operación correcta'
          content:
            application/json:
              schema:
                type: object
                required: [success, data]
                properties:
                  success:
                    const: true
                  data: { $ref: '#/components/schemas/InvitationPreview' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /api/v1/groups/{groupId}/categories:
    parameters:
      - $ref: '#/components/parameters/GroupId'
    get:
      tags: [Groups]
      summary: 'Lista categorías del grupo'
      operationId: listCategories
      security:
        - sessionCookie: []
      responses:
        '200':
          description: 'Operación correcta'
          content:
            application/json:
              schema:
                type: object
                required: [success, data]
                properties:
                  success:
                    const: true
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Category' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
    post:
      tags: [Groups]
      summary: 'Crea una categoría (propietario o administrador)'
      operationId: createCategory
      security:
        - sessionCookie: []
          csrfCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CategoryInput' }
      responses:
        '201':
          description: 'Operación correcta'
          content:
            application/json:
              schema:
                type: object
                required: [success, data]
                properties:
                  success:
                    const: true
                  data: { $ref: '#/components/schemas/Category' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /api/v1/groups/{groupId}/categories/{categoryId}:
    parameters:
      - $ref: '#/components/parameters/GroupId'
      - $ref: '#/components/parameters/CategoryId'
    delete:
      tags: [Groups]
      summary: 'Elimina una categoría; los gastos quedan sin categoría'
      operationId: deleteCategory
      security:
        - sessionCookie: []
          csrfCookie: []
          csrfHeader: []
      responses:
        '200':
          description: 'Operación correcta'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MutationEnvelope' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: cabales_session
      description: Nombre predeterminado configurable mediante COOKIE_NAME.
    csrfCookie:
      type: apiKey
      in: cookie
      name: cabales_session_csrf
      description: Cookie de doble envío cuyo prefijo sigue COOKIE_NAME; /auth/me valida su hash antes de devolverla.
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: Debe coincidir con csrfCookie y con el hash ligado a la sesión.
  parameters:
    GroupId: { name: groupId, in: path, required: true, schema: { type: string, format: uuid } }
...
    InvitationId:
      name: invitationId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    CategoryId:
      name: categoryId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    MutationSuccess:
      {
        description: Mutación completada,
        content:
          { application/json: { schema: { $ref: '#/components/schemas/MutationEnvelope' } } },
      }
    GroupSuccess:
      {
        description: Grupo,
        content: { application/json: { schema: { $ref: '#/components/schemas/GroupEnvelope' } } },
      }
    GroupListSuccess:
      {
        description: Grupos,
        content:
          { application/json: { schema: { $ref: '#/components/schemas/GroupListEnvelope' } } },
      }
    GroupDetailSuccess:
      {
        description: Grupo y miembros,
        content:
          { application/json: { schema: { $ref: '#/components/schemas/GroupDetailEnvelope' } } },
      }
    InvitationSuccess:
      {
        description: Invitación y token efímero,
        content:
          { application/json: { schema: { $ref: '#/components/schemas/InvitationEnvelope' } } },
      }
    MembershipSuccess:
      {
        description: Membresía aceptada,
        content:
          { application/json: { schema: { $ref: '#/components/schemas/MembershipEnvelope' } } },
      }
...
    BadRequest:
      {
        description: Entrada inválida,
        content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } },
      }
    Unauthorized:
      {
        description: Sesión ausente o inválida,
        content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } },
      }
    Forbidden:
      {
        description: CSRF o permiso inválido,
        content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } },
      }
    NotFound:
      {
        description: Recurso no visible,
        content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } },
      }
    Conflict:
      {
        description: Estado concurrente o idempotencia incompatible,
        content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } },
      }
    Gone:
      {
        description: Invitación vencida,
        content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } },
      }
    UnprocessableEntity:
      {
        description: Invariante de negocio incumplida,
        content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } },
      }
    TooManyRequests:
      {
        description: Límite de solicitudes excedido,
        content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } },
      }
...
  schemas:
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [success, error]
      properties:
        success: { const: false }
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: { type: string }
            message: { type: string }
            requestId: { type: string }
            details: {}
...
    MutationEnvelope:
      type: object
      required: [success, data]
      properties:
        {
          success: { const: true },
          data: { type: object, additionalProperties: { type: boolean } },
        }
    User:
      type: object
      additionalProperties: false
      required: [id, email, displayName, avatarUrl, locale, emailVerified]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        displayName: { type: string }
        avatarUrl: { type: [string, 'null'], format: uri }
        locale: { type: string }
        emailVerified: { type: boolean }
...
    Group:
      type: object
      additionalProperties: false
      required: [id, name, description, currency, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        description: { type: [string, 'null'] }
        currency: { type: string, pattern: '^[A-Z]{3}$' }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    GroupEnvelope:
      type: object
      required: [success, data]
      properties: { success: { const: true }, data: { $ref: '#/components/schemas/Group' } }
    GroupListItem:
      allOf:
        - $ref: '#/components/schemas/Group'
        - type: object
          required: [members, _count]
          properties:
            members: { type: array, items: { $ref: '#/components/schemas/Membership' } }
            _count:
              {
                type: object,
                required: [members, events, expenses],
                properties:
                  {
                    members: { type: integer },
                    events: { type: integer },
                    expenses: { type: integer },
                  },
              }
    GroupListEnvelope:
      type: object
      required: [success, data]
      properties:
        {
          success: { const: true },
          data: { type: array, items: { $ref: '#/components/schemas/GroupListItem' } },
        }
    Membership:
      type: object
      required: [id, role]
      properties:
        id: { type: string, format: uuid }
        groupId: { type: string, format: uuid }
        role: { enum: [OWNER, ADMIN, MEMBER] }
        joinedAt: { type: string, format: date-time }
    GroupDetail:
      allOf:
        - $ref: '#/components/schemas/Group'
        - type: object
          required: [members]
          properties:
            members:
              type: array
              items:
                allOf:
                  - $ref: '#/components/schemas/Membership'
                  - type: object
                    required: [user]
                    properties: { user: { $ref: '#/components/schemas/User' } }
    GroupDetailEnvelope:
      type: object
      required: [success, data]
      properties: { success: { const: true }, data: { $ref: '#/components/schemas/GroupDetail' } }
    GroupInput:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name: { type: string, minLength: 2, maxLength: 120 }
        description: { type: string, maxLength: 500 }
        currency: { type: string, pattern: '^[A-Z]{3}$', default: USD }
    GroupPatch:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        name: { type: string, minLength: 2, maxLength: 120 }
        description: { type: [string, 'null'], maxLength: 500 }
        currency: { type: string, pattern: '^[A-Z]{3}$' }
    InvitationInput:
      type: object
      additionalProperties: false
      required: [email]
      properties:
        {
          email: { type: string, format: email, maxLength: 320 },
          role: { enum: [ADMIN, MEMBER], default: MEMBER },
        }
    AcceptInvitationInput:
      type: object
      additionalProperties: false
      required: [token]
      properties: { token: { type: string, minLength: 20, maxLength: 200 } }
    InvitationEnvelope:
      type: object
      required: [success, data]
      properties:
        success: { const: true }
        data:
          type: object
          required: [invitation, token, delivery]
          properties:
            token:
              {
                type: string,
                description: Se muestra una sola vez para compartir el enlace manualmente.,
              }
            delivery:
              { enum: [email, manual], description: manual cuando no hay proveedor de correo real. }
            invitation:
              {
                type: object,
                required: [id, groupId, email, role, status, expiresAt],
                properties:
                  {
                    id: { type: string, format: uuid },
                    groupId: { type: string, format: uuid },
                    email: { type: string, format: email },
                    role: { enum: [ADMIN, MEMBER] },
                    status: { const: PENDING },
                    expiresAt: { type: string, format: date-time },
                  },
              }
    MembershipEnvelope:
      type: object
      required: [success, data]
      properties: { success: { const: true }, data: { $ref: '#/components/schemas/Membership' } }
...
    Invitation:
      type: object
      required:
        [
          id,
          groupId,
          email,
          role,
          status,
          expiresAt,
          createdAt,
          acceptedAt,
          revokedAt,
          lastSentAt,
          sendCount,
          invitedBy,
        ]
      properties:
        id:
          type: string
          format: uuid
        groupId:
          type: string
          format: uuid
        email:
          type: string
          format: email
        role:
          enum: [ADMIN, MEMBER]
        status:
          enum: [PENDING, ACCEPTED, REVOKED, EXPIRED]
        expiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        acceptedAt:
          type: [string, 'null']
          format: date-time
        revokedAt:
          type: [string, 'null']
          format: date-time
        lastSentAt:
          type: [string, 'null']
          format: date-time
        sendCount:
          type: integer
        invitedBy:
          type: object
          required: [id, displayName]
          properties:
            id:
              type: string
              format: uuid
            displayName:
              type: string
    InvitationPreview:
      type: object
      required: [groupName, invitedBy, role, status, expiresAt, emailMatches]
      properties:
        groupName:
          type: string
        invitedBy:
          type: string
        role:
          enum: [ADMIN, MEMBER]
        status:
          enum: [PENDING, ACCEPTED, REVOKED, EXPIRED]
        expiresAt:
          type: string
          format: date-time
        emailMatches:
          type: boolean
    Category:
      type: object
      required: [id, groupId, name, color]
      properties:
        id:
          type: string
          format: uuid
        groupId:
          type: string
          format: uuid
        name:
          type: string
        color:
          type: [string, 'null']
    CategoryInput:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 80
        color:
          type: string
          pattern: '^#[0-9a-fA-F]{6}$'
...